Override find_by_sql instead of init_with
Overriding init_with results in extra cycles burned for every record
instantiated. For the case we're optimizing for, where we're selecting
results with one or more aliased columns, it makes more sense to do the
extra processing *once* and then only include the created module into
each instance.

If no method in this module is ever used, then the actual processing
time will actually increase slightly over the case where we do not
include a module into the singleton class; this is expected since we're
doing extra work. However, if we call any method on the module at least
once per instance, which is typically the use case, then performance is
improved. This improvement is proportional to the number of times we
call the methods on the included module (i.e. column aliases).